477 refactored api, inspect, added name to given, added new format, added inspect api
