+++ spinta/cli/push.py
@@ -39,6 +39,7 @@ from spinta.cli.helpers.data import count_rows
 from spinta.cli.helpers.data import ensure_data_dir
 from spinta.cli.helpers.data import iter_model_rows
 from spinta.cli.helpers.errors import ErrorCounter
+from spinta.cli.helpers.manifest import convert_str_to_manifest_path
 from spinta.cli.helpers.store import prepare_manifest
 from spinta.client import get_access_token
 from spinta.client import get_client_credentials
@@ -123,6 +124,7 @@ def push(
     if stop_time:
         stop_time = toseconds(stop_time)
 
+    manifests = convert_str_to_manifest_path(manifests)
     context = configure_context(ctx.obj, manifests, mode=mode)
     store = prepare_manifest(context)
     config: Config = context.get('config')
