+++ spinta/cli/server.py
@@ -9,6 +9,7 @@ from typer import Option
 from typer import echo
 
 from spinta import commands
+from spinta.cli.helpers.manifest import convert_str_to_manifest_path
 from spinta.cli.helpers.store import load_store
 from spinta.cli.helpers.store import prepare_manifest
 from spinta.components import Mode
@@ -36,6 +37,7 @@ def run(
 
     os.environ['AUTHLIB_INSECURE_TRANSPORT'] = '1'
 
+    manifests = convert_str_to_manifest_path(manifests)
     context = configure_context(ctx.obj, manifests, mode=mode, backend=backend)
     prepare_manifest(context, ensure_config_dir=True)
     app = spinta.api.init(context)
