+++ spinta/cli/show.py
@@ -6,6 +6,7 @@ from typer import Context as TyperContext
 from typer import Option
 from typer import echo
 
+from spinta.cli.helpers.manifest import convert_str_to_manifest_path
 from spinta.cli.helpers.store import prepare_manifest
 from spinta.components import Mode
 from spinta.core.context import configure_context
@@ -20,6 +21,7 @@ def show(
     mode: Mode = Option('internal', help="Mode of backend operation"),
 ):
     """Show manifest as ascii table"""
+    manifests = convert_str_to_manifest_path(manifests)
     context = configure_context(ctx.obj, manifests, mode=mode)
     store = prepare_manifest(context, verbose=False)
     manifest = store.manifest
