+++ spinta/core/config.py
@@ -544,7 +544,7 @@ def parse_resource_args(
     return [resource]
 
 
-def _parse_manifest_path(
+def parse_manifest_path(
     rc: RawConfig,
     path: Union[str, ManifestPath],
 ) -> ManifestPath:
@@ -554,6 +554,11 @@ def _parse_manifest_path(
     from spinta.manifests.helpers import detect_manifest_from_path
     Manifest_ = detect_manifest_from_path(rc, path)
     return ManifestPath(type=Manifest_.type, path=path)
+
+
+def check_if_manifest_valid(rc: RawConfig, manifest: str):
+    names = rc.keys('components', 'manifests')
+    return manifest in names
 
 
 def _get_resource_config(
@@ -584,6 +589,7 @@ def configure_rc(
     mode: Mode = Mode.internal,
     backend: str = None,
     resources: List[ResourceTuple] = None,
+    dataset: str = None,
 ) -> RawConfig:
 
     config: Dict[str, Any] = {}
@@ -618,12 +624,17 @@ def configure_rc(
         if manifests:
             for i, path in enumerate(manifests):
                 manifest_name = f'manifest{i}'
-                manifest = _parse_manifest_path(rc, path)
+                manifest = parse_manifest_path(rc, path)
                 config[f'manifests.{manifest_name}'] = {
                     'type': manifest.type,
                     'path': manifest.path,
                     'file': manifest.file,
                 }
+                if dataset:
+                    config[f'manifests.{manifest_name}']['manifest'] = [{
+                        'type': 'dataset',
+                        'name': dataset
+                    }]
                 sync.append(manifest_name)
 
         if resources:
