+++ spinta/datasets/components.py
@@ -26,6 +26,7 @@ from spinta.types.project import Project
 
 class DatasetGiven:
     access: str = None
+    name: str = None
 
 
 class Dataset(MetaData):
@@ -84,6 +85,7 @@ class Dataset(MetaData):
             }
         },
         'source': {'type': 'string'},
+        'given_name': {'type': 'string', 'default': None},
     }
 
     def __init__(self):
@@ -102,6 +104,7 @@ class External(Node):
 
 class ResourceGiven:
     access: Access = None
+    name: str = None
 
 
 class Resource(External):
@@ -151,6 +154,7 @@ class Resource(External):
         'title': {'type': 'string'},
         'description': {'type': 'string'},
         'comments': {},
+        'given_name': {'type': 'string', 'default': None},
     }
 
     def __init__(self):
